Fix compilation errors after migrating to newer range version

+++ src/algorithms/refinement/ruin/RemoveRandomRoutes.hpp
@@ -20,9 +20,9 @@ struct remove_random_routes final {
   void operator()(const RefinementContext& rCtx,
                   const models::Solution& sln,
                   construction::InsertionContext& iCtx) const {
-    auto max = std::max(static_cast<size_t>(iCtx.solution->routes.size() * threshold), static_cast<size_t>(rmin));
-    auto toDelete = std::min(static_cast<size_t>(rCtx.random->uniform<int>(rmin, rmax)),  //
-                             std::min(max, iCtx.solution->routes.size()));
+    auto max = std::max(static_cast<int>(iCtx.solution->routes.size() * threshold), rmin);
+    auto toDelete = std::min(rCtx.random->uniform<int>(rmin, rmax),  //
+                             std::min(max, static_cast<int>(iCtx.solution->routes.size())));
 
     ranges::for_each(ranges::view::iota(0, toDelete), [&](auto) {
       auto routeIndex = rCtx.random->uniform<int>(0, static_cast<int>(iCtx.solution->routes.size()) - 1);
